+++ lib/woocommerce/woocommerce-notice.php
@@ -7,12 +7,12 @@
  * @package Genesis Sample
  * @author  StudioPress
  * @license GPL-2.0+
- * @link    http://www.studiopress.com/
+ * @link    https://www.studiopress.com/
  */
 
 add_action( 'admin_print_styles', 'genesis_sample_remove_woocommerce_notice' );
 /**
- * Remove the default WooCommerce Notice.
+ * Removes the default WooCommerce Notice.
  *
  * @since 2.3.0
  */
@@ -29,7 +29,7 @@ function genesis_sample_remove_woocommerce_notice() {
 
 add_action( 'admin_notices', 'genesis_sample_woocommerce_theme_notice' );
 /**
- * Add a prompt to activate Genesis Connect for WooCommerce
+ * Adds a prompt to activate Genesis Connect for WooCommerce
  * if WooCommerce is active but Genesis Connect is not.
  *
  * @since 2.3.0
@@ -50,52 +50,61 @@ function genesis_sample_woocommerce_theme_notice() {
 	if ( get_user_option( 'genesis_sample_woocommerce_message_dismissed', get_current_user_id() ) ) {
 		return;
 	}
+
+	/* translators: %s: child theme name */
 	$notice_html = sprintf( __( 'Please install and activate <a href="https://wordpress.org/plugins/genesis-connect-woocommerce/" target="_blank">Genesis Connect for WooCommerce</a> to <strong>enable WooCommerce support for %s</strong>.', 'genesis-sample' ), esc_html( CHILD_THEME_NAME ) );
 
 	if ( current_user_can( 'install_plugins' ) ) {
 		$plugin_slug  = 'genesis-connect-woocommerce';
 		$admin_url    = network_admin_url( 'update.php' );
-		$install_link = sprintf( '<a href="%s">%s</a>', wp_nonce_url(
-			add_query_arg(
-				array(
-					'action' => 'install-plugin',
-					'plugin' => $plugin_slug,
+		$install_link = sprintf(
+			'<a href="%s">%s</a>', wp_nonce_url(
+				add_query_arg(
+					array(
+						'action' => 'install-plugin',
+						'plugin' => $plugin_slug,
+					),
+					$admin_url
 				),
-				$admin_url
-			),
-			'install-plugin_' . $plugin_slug
-		), __( 'install and activate Genesis Connect for WooCommerce', 'genesis-sample' ) );
+				'install-plugin_' . $plugin_slug
+			), __( 'install and activate Genesis Connect for WooCommerce', 'genesis-sample' )
+		);
 
-		$notice_html = sprintf( __( 'Please %s to <strong>enable WooCommerce support for %s</strong>.', 'genesis-sample' ), $install_link, esc_html( CHILD_THEME_NAME ) );
+		/* translators: 1: plugin install prompt presented as link, 2: child theme name */
+		$notice_html = sprintf( __( 'Please %1$s to <strong>enable WooCommerce support for %2$s</strong>.', 'genesis-sample' ), $install_link, esc_html( CHILD_THEME_NAME ) );
 	}
 
-	echo '<div class="notice notice-info is-dismissible genesis-sample-woocommerce-notice"><p>' . $notice_html . '</p></div>';
+	echo '<div class="notice notice-info is-dismissible genesis-sample-woocommerce-notice"><p>' . wp_kses_post( $notice_html ) . '</p></div>';
 
 }
 
 add_action( 'wp_ajax_genesis_sample_dismiss_woocommerce_notice', 'genesis_sample_dismiss_woocommerce_notice' );
 /**
- * Add option to dismiss Genesis Connect for Woocommerce plugin install prompt.
+ * Adds option to dismiss Genesis Connect for Woocommerce plugin install prompt.
  *
  * @since 2.3.0
  */
 function genesis_sample_dismiss_woocommerce_notice() {
+
 	update_user_option( get_current_user_id(), 'genesis_sample_woocommerce_message_dismissed', 1 );
+
 }
 
 add_action( 'admin_enqueue_scripts', 'genesis_sample_notice_script' );
 /**
- * Enqueue script to clear the Genesis Connect for WooCommerce plugin install prompt on dismissal.
+ * Enqueues script to clear the Genesis Connect for WooCommerce plugin install prompt on dismissal.
  *
  * @since 2.3.0
  */
 function genesis_sample_notice_script() {
-	wp_enqueue_script( 'genesis_sample_notice_script', get_stylesheet_directory_uri() . '/lib/woocommerce/js/notice-update.js', array( 'jquery' ), '1.0', true  );
+
+	wp_enqueue_script( 'genesis_sample_notice_script', get_stylesheet_directory_uri() . '/lib/woocommerce/js/notice-update.js', array( 'jquery' ), '1.0', true );
+
 }
 
 add_action( 'switch_theme', 'genesis_sample_reset_woocommerce_notice', 10, 2 );
 /**
- * Clear the Genesis Connect for WooCommerce plugin install prompt on theme change.
+ * Clears the Genesis Connect for WooCommerce plugin install prompt on theme change.
  *
  * @since 2.3.0
  */
@@ -103,7 +112,7 @@ function genesis_sample_reset_woocommerce_notice() {
 
 	global $wpdb;
 
-	$args = array(
+	$args  = array(
 		'meta_key'   => $wpdb->prefix . 'genesis_sample_woocommerce_message_dismissed',
 		'meta_value' => 1,
 	);
@@ -117,17 +126,17 @@ function genesis_sample_reset_woocommerce_notice() {
 
 add_action( 'deactivated_plugin', 'genesis_sample_reset_woocommerce_notice_on_deactivation', 10, 2 );
 /**
- * Clear the Genesis Connect for WooCommerce plugin prompt on deactivation.
+ * Clears the Genesis Connect for WooCommerce plugin prompt on deactivation.
  *
  * @since 2.3.0
  *
  * @param string $plugin The plugin slug.
- * @param $network_activation.
+ * @param bool   $network_deactivating Whether the plugin is deactivated for all sites in the network. or just the current site.
  */
-function genesis_sample_reset_woocommerce_notice_on_deactivation( $plugin, $network_activation ) {
+function genesis_sample_reset_woocommerce_notice_on_deactivation( $plugin, $network_deactivating ) {
 
 	// Conditional check to see if we're deactivating WooCommerce or Genesis Connect for WooCommerce.
-	if ( $plugin !== 'woocommerce/woocommerce.php' && $plugin !== 'genesis-connect-woocommerce/genesis-connect-woocommerce.php'  ) {
+	if ( 'woocommerce/woocommerce.php' !== $plugin && 'genesis-connect-woocommerce/genesis-connect-woocommerce.php' !== $plugin ) {
 		return;
 	}