+++ lib/woocommerce/woocommerce-setup.php
@@ -7,12 +7,21 @@
  * @package Genesis Sample
  * @author  StudioPress
  * @license GPL-2.0+
- * @link    http://www.studiopress.com/
+ * @link    https://www.studiopress.com/
  */
+
+// Adds product gallery support.
+if ( class_exists( 'WooCommerce' ) ) {
+
+	add_theme_support( 'wc-product-gallery-lightbox' );
+	add_theme_support( 'wc-product-gallery-slider' );
+	add_theme_support( 'wc-product-gallery-zoom' );
+
+}
 
 add_action( 'wp_enqueue_scripts', 'genesis_sample_products_match_height', 99 );
 /**
- * Print an inline script to the footer to keep products the same height.
+ * Prints an inline script to the footer to keep products the same height.
  *
  * @since 2.3.0
  */
@@ -23,14 +32,23 @@ function genesis_sample_products_match_height() {
 		return;
 	}
 
-	wp_enqueue_script( 'genesis-sample-match-height', get_stylesheet_directory_uri() . '/js/jquery.matchHeight.min.js', array( 'jquery' ), CHILD_THEME_VERSION, true );
-	wp_add_inline_script( 'genesis-sample-match-height', "jQuery(document).ready( function() { jQuery( '.product .woocommerce-LoopProduct-link').matchHeight(); });" );
+	wp_enqueue_script(
+		'genesis-sample-match-height',
+		get_stylesheet_directory_uri() . '/js/jquery.matchHeight.min.js',
+		array( 'jquery' ),
+		CHILD_THEME_VERSION,
+		true
+	);
+	wp_add_inline_script(
+		'genesis-sample-match-height',
+		"jQuery(document).ready( function() { jQuery( '.product .woocommerce-LoopProduct-link').matchHeight(); });"
+	);
 
 }
 
 add_filter( 'woocommerce_style_smallscreen_breakpoint', 'genesis_sample_woocommerce_breakpoint' );
 /**
- * Modify the WooCommerce breakpoints.
+ * Modifies the WooCommerce breakpoints.
  *
  * @since 2.3.0
  *
@@ -44,44 +62,38 @@ function genesis_sample_woocommerce_breakpoint() {
 			'content-sidebar',
 			'sidebar-content',
 		),
-		'two-sidebar' => array(
-			'content-sidebar-sidebar',
-			'sidebar-content-sidebar',
-			'sidebar-sidebar-content',
-		),
 	);
 
-	if ( in_array( $current, $layouts['two-sidebar'] ) ) {
-		return '2000px'; // Show mobile styles immediately.
-	}
-	elseif ( in_array( $current, $layouts['one-sidebar'] ) ) {
+	if ( in_array( $current, $layouts['one-sidebar'], true ) ) {
 		return '1200px';
 	}
-	else {
-		return '860px';
-	}
+
+	return '860px';
 
 }
 
 add_filter( 'genesiswooc_products_per_page', 'genesis_sample_default_products_per_page' );
 /**
- * Set the default products per page.
+ * Sets the default products per page.
  *
  * @since 2.3.0
  *
  * @return int Number of products to show per page.
  */
 function genesis_sample_default_products_per_page() {
+
 	return 8;
+
 }
 
-add_filter( 'woocommerce_pagination_args', 	'genesis_sample_woocommerce_pagination' );
+add_filter( 'woocommerce_pagination_args', 'genesis_sample_woocommerce_pagination' );
 /**
- * Update the next and previous arrows to the default Genesis style.
+ * Updates the next and previous arrows to the default Genesis style.
  *
+ * @param array $args The previous and next text arguments.
  * @since 2.3.0
  *
- * @return string New next and previous text string.
+ * @return array New next and previous text arguments.
  */
 function genesis_sample_woocommerce_pagination( $args ) {
 
@@ -94,15 +106,16 @@ function genesis_sample_woocommerce_pagination( $args ) {
 
 add_action( 'after_switch_theme', 'genesis_sample_woocommerce_image_dimensions_after_theme_setup', 1 );
 /**
-* Define WooCommerce image sizes on theme activation.
-*
-* @since 2.3.0
-*/
+ * Defines WooCommerce image sizes on theme activation.
+ *
+ * @since 2.3.0
+ */
 function genesis_sample_woocommerce_image_dimensions_after_theme_setup() {
 
 	global $pagenow;
 
-	if ( ! isset( $_GET['activated'] ) || $pagenow != 'themes.php' || ! class_exists( 'WooCommerce' ) ) {
+	// Checks conditionally to see if we're activating the current theme and that WooCommerce is installed.
+	if ( ! isset( $_GET['activated'] ) || 'themes.php' !== $pagenow || ! class_exists( 'WooCommerce' ) ) { // phpcs:ignore WordPress.CSRF.NonceVerification.NoNonceVerification -- low risk, follows official snippet at https://goo.gl/nnHHQa.
 		return;
 	}
 
@@ -112,14 +125,16 @@ function genesis_sample_woocommerce_image_dimensions_after_theme_setup() {
 
 add_action( 'activated_plugin', 'genesis_sample_woocommerce_image_dimensions_after_woo_activation', 10, 2 );
 /**
- * Define the WooCommerce image sizes on WooCommerce activation.
+ * Defines the WooCommerce image sizes on WooCommerce activation.
  *
  * @since 2.3.0
+ *
+ * @param string $plugin The path of the plugin being activated.
  */
 function genesis_sample_woocommerce_image_dimensions_after_woo_activation( $plugin ) {
 
-	// Check to see if WooCommerce is being activated.
-	if ( $plugin !== 'woocommerce/woocommerce.php' ) {
+	// Checks to see if WooCommerce is being activated.
+	if ( 'woocommerce/woocommerce.php' !== $plugin ) {
 		return;
 	}
 
@@ -128,31 +143,38 @@ function genesis_sample_woocommerce_image_dimensions_after_woo_activation( $plug
 }
 
 /**
- * Update WooCommerce image dimensions.
+ * Updates WooCommerce image dimensions.
  *
  * @since 2.3.0
  */
 function genesis_sample_update_woocommerce_image_dimensions() {
 
-	$catalog = array(
-		'width'  => '500', // px
-		'height' => '500', // px
-		'crop'   => 1,     // true
-	);
-	$single = array(
-		'width'  => '655', // px
-		'height' => '655', // px
-		'crop'   => 1,     // true
-	);
-	$thumbnail = array(
-		'width'  => '180', // px
-		'height' => '180', // px
-		'crop'   => 1,     // true
+	// Updates image size options.
+	update_option( 'woocommerce_single_image_width', 655 );    // Single product image.
+	update_option( 'woocommerce_thumbnail_image_width', 500 ); // Catalog image.
+
+	// Updates image cropping option.
+	update_option( 'woocommerce_thumbnail_cropping', '1:1' );
+
+}
+
+add_filter( 'woocommerce_get_image_size_gallery_thumbnail', 'genesis_sample_gallery_image_thumbnail' );
+/**
+ * Filters the WooCommerce gallery image dimensions.
+ *
+ * @since 2.6.0
+ *
+ * @param array $size The gallery image size and crop arguments.
+ * @return array The modified gallery image size and crop arguments.
+ */
+function genesis_sample_gallery_image_thumbnail( $size ) {
+
+	$size = array(
+		'width'  => 180,
+		'height' => 180,
+		'crop'   => 1,
 	);
 
-	// Image sizes.
-	update_option( 'shop_catalog_image_size', $catalog );     // Product category thumbs.
-	update_option( 'shop_single_image_size', $single );       // Single product image.
-	update_option( 'shop_thumbnail_image_size', $thumbnail ); // Image gallery thumbs.
+	return $size;
 
 }